pyproject.toml
[tool.poetry]
name = "weirdsheeplabs-cdn"
version = "0.1.0"
description = "CDN for weirdsheeplabs.com"
authors = ["Armand Rego <armand@weirdsheeplabs.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aws-cdk-lib = "^2.132.0"
constructs = "^10.3.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
